refactor(bootstrap): extract path resolution to layer/paths
Phase 4.5b of the uc-bootstrap redesign. Move the four directory-layout
functions (get_default_app_dirs, get_storage_paths, resolve_app_paths,
apply_profile_suffix) out of assembly.rs into layer/paths.rs, establishing
the layer/ subdir. get_storage_paths re-exports from layer::paths; internal
callers updated; assembly.rs imports the three it still uses. assembly.rs:
1345 -> 1258 lines. No behavior change.
