+++ crates/uc-bootstrap/src/entrypoint/non_gui.rs
@@ -41,7 +41,7 @@ use uc_infra::mobile_sync::{
 };
 use uc_infra::network::iroh::{IrohRelayProbeAdapter, IrohRelayProbeError, IrohRelayProbeReport};
 
-use crate::assembly::get_storage_paths;
+use crate::layer::paths::get_storage_paths;
 use crate::subsystem::sync_engine::{build_sync_engine_assembly, SyncEngineAssembly};
 
 // ---------------------------------------------------------------------------
