+++ crates/uc-bootstrap/src/lib.rs
@@ -7,6 +7,7 @@
 
 pub mod assembly;
 pub mod entrypoint;
+pub mod layer;
 mod network_policy;
 pub mod observability;
 pub mod startup;
@@ -22,13 +23,14 @@ pub mod wiring;
 // `compose_event_context` 在 `wire_dependencies` 之后由各进程入口调用一次。
 pub use subsystem::analytics::compose_event_context;
 
-pub use assembly::{get_storage_paths, wire_dependencies, SystemClipboardWiring};
+pub use assembly::{wire_dependencies, SystemClipboardWiring};
 pub use entrypoint::daemon::{build_daemon_lifecycle, DaemonLifecycle};
 pub use entrypoint::non_gui::{
     build_app_facade_from_deps, build_cli_app_runtime, build_mobile_sync_facade,
     resolve_clipboard_integration_mode, AppFacadeAssemblyOptions, CliAppRuntime,
     ClipboardRestoreAssembly,
 };
+pub use layer::paths::get_storage_paths;
 pub use observability::tracing::{init_tracing_subscriber, install_panic_logging_hook};
 pub use subsystem::blob_tasks::{spawn_blob_processing_tasks, BlobProcessingPorts};
 pub use subsystem::file_transfer::FileTransferLifecycle;
